requirements.txt
fastapi~=0.110.0
SQLAlchemy~=2.0.27
python-dotenv~=1.0.1
uvicorn~=0.27.1
